+++ src/class-wc-gateway-bitpay.php
@@ -1,1 +1,17 @@
-<?php require_once __DIR__ ."/class-wc-gateway-btcpay.php" ?>
+<?php 
+/*
+    Plugin Name: BTCPay for WooCommerce
+    Plugin URI:  https://wordpress.org/plugins/btcpay-for-woocommerce
+    Description: Enable your WooCommerce store to accept Bitcoin with BTCPay.
+    Author:      BTCPay
+    Text Domain: BTCPay
+    Author URI:  https://github.com/btcpayserver
+
+    Version:           3.0.4
+    License:           Copyright 2011-2018 BTCPay & BitPay Inc., MIT License
+    License URI:       https://github.com/btcpayserver/woocommerce-plugin/blob/master/LICENSE
+    GitHub Plugin URI: https://github.com/btcpayserver/woocommerce-plugin
+*/
+
+    require_once __DIR__ ."/class-wc-gateway-btcpay.php" 
+?>